Refactor: Move protobuf contracts and mappers for business process requests to BusinessRequestProcessing namespace to improve overview and maintainability
