+++ source/Energinet.DataHub.MarketRoles.Infrastructure/BusinessRequestProcessing/Protobuf/Contracts/Contract.proto
